[refactor] OpenAIClient 구조 개선 및 응답 DTO 분리
